feat: separando responsabilidades da tela de login
