+++ src/pages/Login.jsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/components/login-form'
+import { LoginForm } from '@/components/LoginForm'
 import loginBg from '../assets/images/login-bg.jpg'
 import logoWhite from '../assets/images/logo-white.png'
 import logo from '../assets/images/logo.png'
@@ -7,7 +7,6 @@ import React from 'react'
 const Login = () => {
   return (
     <div className="flex min-h-screen">
-      {/* left side */}
       <div className="relative z-0 hidden w-1/2 flex-col justify-between bg-zinc-900 p-15 text-white lg:flex">
         <img
           src={loginBg}
@@ -28,7 +27,6 @@ const Login = () => {
         </div>
       </div>
 
-      {/* right side */}
       <div className="flex w-full items-center justify-center bg-white lg:w-1/2">
         <div className="w-full max-w-120 space-y-8 px-6">
           <div className="m-0 flex justify-center">
